+++ src/PhpSpreadsheet/Writer/Xlsx/Chart.php
@@ -1098,6 +1098,20 @@ class Chart extends WriterPart
         foreach ($plotSeriesOrder as $plotSeriesIdx => $plotSeriesRef) {
             $objWriter->startElement('c:ser');
 
+            $plotLabel = $plotGroup->getPlotLabelByIndex($plotSeriesIdx);
+            if ($plotLabel) {
+                $fillColor = $plotLabel->getFillColor();
+                if ($fillColor !== null) {
+                    $objWriter->startElement('c:spPr');
+                    $objWriter->startElement('a:solidFill');
+                    $objWriter->startElement('a:srgbClr');
+                    $objWriter->writeAttribute('val', $fillColor);
+                    $objWriter->endElement();
+                    $objWriter->endElement();
+                    $objWriter->endElement();
+                }
+            }
+
             $objWriter->startElement('c:idx');
             $objWriter->writeAttribute('val', $this->_seriesIndex + $plotSeriesIdx);
             $objWriter->endElement();